+++ config/protocol.php
@@ -1,0 +1,14 @@
+<?php
+
+// config for EdStevo/Standards
+return [
+    /*
+    |--------------------------------------------------------------------------
+    | EdStevo Standards Configuration
+    |--------------------------------------------------------------------------
+    |
+    | This configuration file allows you to customize the behavior of the
+    | EdStevo Standards package.
+    |
+    */
+];